Persistence of selected genes and regulons
Selected genes and regulons persist when zooming, sorting and clustering the display. They can also be saved and loaded in CSVs (though due to how genes are selected/highlighted on hover in this version of TRNDiff, if a CSV has at least one gene of a certain name selected, all instances of that gene will be selected after loading). To help facilitate this, the selected gene names and regulon IDs are stored in arrays.

Another slight internal change was to store only the index of the dragged wagon wheel rather than (a reference to) the whole element object.

Also cleaned up the code slightly (at least for the parts I've worked on) - all instances of "==" and "!=" have been replaced with "===" and "!==", and any "null" that was used with the former equality operators was changed to "undefined". Mentions of "graph" in wagon-wheel.js have been changed to "wheel" for consistency - wagon-wheel objects were referred to as "graphs" in the code of older version of TRNDiff and that was what I originally used in my additions, but as Joe used "(wagon) wheel", I thought it would be prudent to change them.