Scrolling disabled while dragging a graph
When dragging a graph, causing the page to scroll will cause odd results. For this reason, scrolling is disabled while a pan that causes a graph to be dragged, and enabled again once the action is complete or cancelled.

There may be instances where a user would want to drag a graph to a position that is not visible in the current scroll position; I'm not sure if this would be trivial to handle, however.